requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.7
pytest==7.4.2
pytest-flask==1.2.0
mysql-connector-python==8.2.0
pymysql==1.1.0